django/requirements.txt
django-cms<3.3
django-treebeard>=2.0
psycopg2
https://github.com/divio/djangocms-column/archive/master.zip?1450269951.83
https://github.com/divio/djangocms-flash/archive/master.zip?1450269951.83
https://github.com/divio/djangocms-googlemap/archive/master.zip?1450269951.83
https://github.com/divio/djangocms-inherit/archive/master.zip?1450269951.83
https://github.com/divio/djangocms-link/archive/master.zip?1450269951.83
https://github.com/divio/djangocms-style/archive/master.zip?1450269951.83
easy_thumbnails
django-filer>=0.9.12
cmsplugin-filer>=0.10.2
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.0
Django==1.8.6
pytz
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six
django-post_office==2.0.3
django-smuggler
django-widget-tweaks
django-storages
boto
kombu
celery
django-celery
